LRM FCS nerfed
So, I know a few of you lurk the Git so let me explain: there is no *weapon field* in CAC or vanilla that I can see for buffing only Indirect Fire attacks. However, there's a statusEffect, so what I've done is stapled that statusEffect, tagged for LRM-type missiles only, to the FCS item itself. It's dirty but it works for the job.